moved simulate into a separate binary
