+++ src/main.rs
@@ -6,7 +6,9 @@ use std::{
 };
 
 use calypso::{
-    command_data, data::{DecodeData, EncodeData}, decodable_message::DecodableMessage, encodable_message::EncodableMessage, encode_master_mapping::ENCODABLE_KEY_LIST, mqtt::MqttClient, serverdata, simulatable_message::SimulatedComponents, simulate_data::create_simulated_components
+    command_data, data::DecodeData, data::EncodeData, decodable_message::DecodableMessage,
+    encodable_message::EncodableMessage, encode_master_mapping::ENCODABLE_KEY_LIST,
+    mqtt::MqttClient, serverdata,
 };
 use clap::Parser;
 use protobuf::Message;
@@ -21,10 +23,6 @@ struct CalypsoArgs {
     /// Whether to enable CAN message encoding
     #[arg(short = 'e', long, env = "CALYPSO_CAN_ENCODE")]
     encode: bool,
-
-    /// Whether to enable SIMULATION mode
-    #[arg(short = 's', long, env = "CALYPSO_CAN_SIMULATE")]
-    simulate: bool,
 
     /// The host url of the siren, including port and excluding protocol prefix
     #[arg(
@@ -147,40 +145,6 @@ fn read_can(pub_path: &str, can_interface: &str) -> JoinHandle<u32> {
     })
 }
 
-
-fn simulate_out(pub_path: &str) {
-    let mut client = MqttClient::new(pub_path, "calypso-simulator");
-    let _ = client.connect(); // todo: add error handling
-    let sleep_time = Duration::from_millis(10);
-
-    // todo: a way to turn individual components on and off
-    let mut simulated_components: Vec<SimulatedComponents> = create_simulated_components();
-
-    loop {
-        for component in simulated_components.iter_mut() {
-            if component.should_update() {
-                component.update();
-                let data: calypso::data::DecodeData = component.get_data();
-                let mut payload = serverdata::ServerData::new();
-                payload.unit = data.unit.to_string();
-                payload.value = data.value.iter().map(|x| x.to_string()).collect();
-
-                client
-                    .publish(
-                        data.topic.to_string(),
-                        protobuf::Message::write_to_bytes(&payload).unwrap_or_else(|e| {
-                            format!("failed to serialize {}", e).as_bytes().to_vec()
-                        }),
-                    )
-                    .expect("Could not publish!");
-            }
-            // sleep for a bit
-            thread::sleep(sleep_time);
-        }
-    }
-}
-
-
 /**
  * Reads the mqtt incoming messages and sends can messages based off of that
  */
@@ -300,16 +264,6 @@ fn send_out(
  */
 fn main() {
     let cli = CalypsoArgs::parse();
-    
-    if cli.simulate {
-        println!("> Starting simulation mode");
-        simulate_out(&cli.siren_host_url);
-
-        // simulator_handle.join().expect("Simulator failed with ");
-        return;
-    }
-
-
     let can_handle = read_can(&cli.siren_host_url, &cli.socketcan_iface);
 
     // use a arc for mutlithread, and a rwlock to enforce one writer
